# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Lennox Hale Publishing Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /series:
    get:
      operationId: listSeries
      tags: [series]
      summary: List all series
      parameters:
        - name: authorId
          in: query
          required: false
          schema:
            type: integer
        - name: language
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
    post:
      operationId: createSeries
      tags: [series]
      summary: Create a new series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
  /series/{id}:
    get:
      operationId: getSeries
      tags: [series]
      summary: Get series by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesWithBooks"
        "404":
          description: Not found
    put:
      operationId: updateSeries
      tags: [series]
      summary: Update a series
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
    delete:
      operationId: deleteSeries
      tags: [series]
      summary: Delete a series
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Books ──
...
components:
  schemas:
...
    Series:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        language:
          type: string
        status:
          type: string
          enum: [active, planned, completed]
        displayOrder:
          type: integer
        crossoverFromSeriesId:
          type: integer
          nullable: true
        crossoverFromSeriesName:
          type: string
          nullable: true
        authorPenName:
          type: string
        createdAt:
          type: string
          format: date-time
        bookCount:
          type: integer
      required: [id, authorId, name, language, status, displayOrder, authorPenName, createdAt, bookCount]
    SeriesWithBooks:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        language:
          type: string
        status:
          type: string
          enum: [active, planned, completed]
        displayOrder:
          type: integer
        crossoverFromSeriesId:
          type: integer
          nullable: true
        crossoverFromSeriesName:
          type: string
          nullable: true
        authorPenName:
          type: string
        createdAt:
          type: string
          format: date-time
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
      required: [id, authorId, name, language, status, displayOrder, authorPenName, createdAt, books]
    CreateSeriesBody:
      type: object
      properties:
        authorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        language:
          type: string
        status:
          type: string
          enum: [active, planned, completed]
        displayOrder:
          type: integer
        crossoverFromSeriesId:
          type: integer
          nullable: true
      required: [authorId, name]
    Book:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        bookNumber:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        language:
          type: string
        wordCount:
          type: integer
          nullable: true
        funnelRole:
          type: string
          enum: [lead_magnet, traffic_entry, core_offer, crossover_bridge]
        pricingStrategy:
          type: string
          enum: [perma_free, promotional, full_price]
        price:
          type: number
          nullable: true
        promotionalPrice:
          type: number
          nullable: true
        status:
          type: string
          enum: [draft, production, ready, scheduled, published]
        publicationDate:
          type: string
          format: date-time
          nullable: true
        scheduledDate:
          type: string
          format: date-time
          nullable: true
        distributionChannel:
          type: string
          enum: [kdp, email_exclusive, wide]
          nullable: true
        asin:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        crossoverToSeriesId:
          type: integer
          nullable: true
        seriesName:
          type: string
        authorPenName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, seriesId, bookNumber, title, language, funnelRole, pricingStrategy, status, seriesName, authorPenName, createdAt]
...